Get App data

Get App data

Get App data

Details

R6 class that allows to get the information needed for an User session of the app.

Super classes

reportesAPI::DBManager -> reportesAPI::User -> reportesAPI::Organisation -> reportesAPI::Group -> reportesAPI::Task -> AppData

Methods

Public methods

AppData$new()

AppData$clone()

Inherited methods

reportesAPI::DBManager$db_execute_statement()

reportesAPI::DBManager$db_get_query()

reportesAPI::DBManager$db_make_query()

reportesAPI::User$user_add()

reportesAPI::User$user_delete()

reportesAPI::User$user_edit_names()

reportesAPI::Organisation$org_add()

reportesAPI::Organisation$org_delete()

reportesAPI::Organisation$org_edit()

reportesAPI::Organisation$org_finalize()

reportesAPI::Organisation$org_initialize()

reportesAPI::Organisation$org_user_add()

reportesAPI::Organisation$org_user_delete()

reportesAPI::Organisation$org_user_edit()

reportesAPI::Group$group_add()

reportesAPI::Group$group_delete()

reportesAPI::Group$group_edit()

reportesAPI::Group$group_initialize()

reportesAPI::Group$group_select()

reportesAPI::Group$group_unit_add()

reportesAPI::Group$group_unit_delete()

reportesAPI::Group$group_unit_edit()

reportesAPI::Group$group_user_add()

reportesAPI::Group$group_user_delete()

reportesAPI::Group$group_user_edit()

reportesAPI::Task$progress_add()

reportesAPI::Task$task_add()

reportesAPI::Task$task_delete()

reportesAPI::Task$task_edit_metadata()

reportesAPI::Task$task_get_history()

reportesAPI::Task$task_report_progress()

Method new()

Start a session of app data based on an user email

Usage

AppData$new(email = Sys.getenv("REPORTES_EMAIL"))

Arguments

email

The email the user started the session with.

Method clone()

The objects of this class are cloneable with this method.

Usage

AppData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.