Get Organisation data

Get Organisation data

Get Organisation data

Details

R6 class that allows to get the Organisation information.

Super classes

reportesAPI::DBManager -> reportesAPI::User -> Organisation

Active bindings

orgs

List containing the organisation affiliations of the User

org_users

List containing the user list of the organisation. The info is shown following the User's organisation role.

Methods

Public methods

Organisation$new()

Organisation$org_initialize()

Organisation$org_edit()

Organisation$org_user_add()

Organisation$org_user_delete()

Organisation$org_user_edit()

Organisation$org_finalize()

Organisation$clone()

Inherited methods

reportesAPI::DBManager$db_execute_statement()

reportesAPI::DBManager$db_get_query()

reportesAPI::DBManager$db_make_query()

reportesAPI::DBManager$get_timestamp()

reportesAPI::User$user_edit_names()

Method new()

Start an Organisation based on an user email

Usage

Organisation$new(email)

Arguments

email

The email the user started the session with.

Method org_initialize()

Initialize an organisation for a new user

Usage

Organisation$org_initialize()

Method org_edit()

Edit Organisation metadata

Usage

Organisation$org_edit(org_id, org_title, org_description)

Arguments

org_id

The id of the organisation on which the statement will be executed

org_title

The new title of the organisation

org_description

The new description of the organisation

Method org_user_add()

Add an user to an organisation

Usage

Organisation$org_user_add(org_id, user_id, org_role)

Arguments

org_id

The id of the organisation on which the statement will be executed

user_id

The id of the user on which the statement will be executed

org_role

The role for the user in the organisation

Method org_user_delete()

Delete an user from an organisation

Usage

Organisation$org_user_delete(org_id, user_id)

Arguments

org_id

The id of the organisation on which the statement will be executed

user_id

The id of the user on which the statement will be executed

Method org_user_edit()

Edit the role of a user inside an organisation

Usage

Organisation$org_user_edit(org_id, user_id, org_role)

Arguments

org_id

The id of the organisation on which the statement will be executed

user_id

The id of the user on which the statement will be executed

org_role

The role for the user in the organisation

Method org_finalize()

Remove the existence of an organisation

Usage

Organisation$org_finalize(org_id)

Arguments

org_id

The id of the organisation on which the statement will be executed

Method clone()

The objects of this class are cloneable with this method.

Usage

Organisation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.