Get Group data

Get Group data

Get Group data

Details

R6 class that allows to get the Group information.

Super classes

reportesAPI::DBManager -> reportesAPI::User -> reportesAPI::Organisation -> reportesAPI::Group -> Task

Active bindings

tasks

List containing the tasks an User can interact with

Methods

Public methods

Task$new()

Task$task_add()

Task$task_delete()

Task$task_report_progress()

Task$task_edit_metadata()

Task$progress_add()

Task$task_get_history()

Task$clone()

Inherited methods

reportesAPI::DBManager$db_execute_statement()

reportesAPI::DBManager$db_get_query()

reportesAPI::DBManager$db_make_query()

reportesAPI::DBManager$get_timestamp()

reportesAPI::User$user_edit_names()

reportesAPI::Organisation$org_edit()

reportesAPI::Organisation$org_finalize()

reportesAPI::Organisation$org_initialize()

reportesAPI::Organisation$org_user_add()

reportesAPI::Organisation$org_user_delete()

reportesAPI::Organisation$org_user_edit()

reportesAPI::Group$group_edit()

reportesAPI::Group$group_finalize()

reportesAPI::Group$group_initialize()

reportesAPI::Group$group_select()

reportesAPI::Group$group_user_add()

reportesAPI::Group$group_user_delete()

reportesAPI::Group$group_user_edit()

Method new()

Start a Task based on an user email

Usage

Task$new(email)

Arguments

email

The email the user started the session with.

Method task_add()

Add a new task for an User

Usage

Task$task_add(
  process_id = NA_character_,
  activity_id = NA_character_,
  org_id,
  group_id,
  task_title,
  task_description,
  assignee,
  time_due,
  output_unit,
  output_goal
)

Arguments

process_id

The id of the process the task belongs to. When used must be paired with activity_id

activity_id

The id of the activity the task belongs to. When used must be paired with process_id

org_id

The id of the organisation on which the statement will be executed

group_id

The id of the group on which the statement will be executed

task_title

The new title of the group

task_description

The new description of the group

assignee

The id of the user responsible for the task

time_due

Deadline for the task completion. Datetime

output_unit

Unit of measurement of the task output

output_goal

Number on which which the output will be measured

Method task_delete()

Delete an user task

Usage

Task$task_delete(
  process_id = NA_character_,
  activity_id = NA_character_,
  org_id,
  group_id,
  task_id
)

Arguments

process_id

The id of the process the task belongs to. When used must be paired with activity_id

activity_id

The id of the activity the task belongs to. When used must be paired with process_id

org_id

The id of the organisation on which the statement will be executed

group_id

The id of the group on which the statement will be executed

task_id

The id of the task on which the statement will be executed

Method task_report_progress()

Report progress on an assigned task

Usage

Task$task_report_progress(
  process_id = NA_character_,
  activity_id = NA_character_,
  org_id,
  group_id,
  task_id,
  status_current,
  output_current,
  details
)

Arguments

process_id

The id of the process the task belongs to. When used must be paired with activity_id

activity_id

The id of the activity the task belongs to. When used must be paired with process_id

org_id

The id of the organisation on which the statement will be executed

group_id

The id of the group on which the statement will be executed

task_id

The id of the task on which the statement will be executed

status_current

Current status of the specified task

output_current

Current output of the specified task

details

Explanation of the progress made

Method task_edit_metadata()

Edit a task metadata

Usage

Task$task_edit_metadata(
  process_id = NA_character_,
  activity_id = NA_character_,
  org_id,
  group_id,
  task_id,
  task_title,
  task_description
)

Arguments

process_id

The id of the process the task belongs to. When used must be paired with activity_id

activity_id

The id of the activity the task belongs to. When used must be paired with process_id

org_id

The id of the organisation on which the statement will be executed

group_id

The id of the group on which the statement will be executed

task_id

The id of the task on which the statement will be executed

task_title

The new title of the group

task_description

The new description of the group

Method progress_add()

Insert progress info on some task

Usage

Task$progress_add(
  process_id = NA_character_,
  activity_id = NA_character_,
  org_id,
  group_id,
  task_id,
  output_progress,
  status,
  details
)

Arguments

process_id

The id of the process the task belongs to. When used must be paired with activity_id

activity_id

The id of the activity the task belongs to. When used must be paired with process_id

org_id

The id of the organisation on which the statement will be executed

group_id

The id of the group on which the statement will be executed

task_id

The id of the task on which the statement will be executed

output_progress

Quantity of the progress being reported. Is measured in the unit specified in the creation of the task

status

The status of the task once the new progress is added

details

Explanation of the progress made

Method task_get_history()

Get a task's progression history

Usage

Task$task_get_history(task_id)

Arguments

task_id

The id of the task on which the statement will be executed

Method clone()

The objects of this class are cloneable with this method.

Usage

Task$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.