Get user information

Get user information

Get user information

Details

R6 class that allows to get the user information.

Super class

reportesAPI::DBManager -> User

Public fields

user

List containing the user info.

Methods

Public methods

User$new()

User$user_add()

User$user_delete()

User$user_edit_names()

User$clone()

Inherited methods

reportesAPI::DBManager$db_execute_statement()

reportesAPI::DBManager$db_get_query()

reportesAPI::DBManager$db_make_query()

Method new()

Start User

Usage

User$new(email)

Arguments

email

The email the user started the session with.

Method user_add()

Add a new user to the database

Usage

User$user_add(name, last_name, email)

Arguments

name

Name to be inserted as user metadata.

last_name

Last name to be inserted as user metadata.

email

The email the user started the session with.

Method user_delete()

Delete a user from the database

Usage

User$user_delete(user_id)

Arguments

user_id

ID of the user to edit or delete.

Method user_edit_names()

Change the name and last name of the User

Usage

User$user_edit_names(user_id, name, last_name)

Arguments

user_id

ID of the user to edit or delete.

name

Name to be inserted as user metadata.

last_name

Last name to be inserted as user metadata.

Method clone()

The objects of this class are cloneable with this method.

Usage

User$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.